Lots of changes to trace, display disconnected nodes, and more.
- Now will display disconnected nodes. The text has a slight
opacity, and there is something in the tooltip to show it's
disconnected for sure.
- If you click on a trace link, it will now move you to the node
shown in path. If that node is hidden, it will unhide it. If that
node isn't listed, it will list it. etc.
- If you toggle a node's children's visibility, and the node moves
off screen, the visualizer now moves so it's still on screen.
- Cache window width and height, avoid relayout as much as possible.
- Add the trace types to the legend.
- If a node is listed and closed right after, add a delay so the
trace is still visible.
- More, I can't remember.